+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.51.3
 - [🚀] 增加`Rect.scale`扩展;
 - [🚀] 增加`SliverStack`;
+- [🚀] `DashedLine`增加`strokeCap`参数;
 
 ## 0.51.2
 - [🚀] `FutureX`增加`thenIf`和`thenIfNot`扩展;